Add files via upload
Make a program to input a certain number of student's information such as name, nim, gender. The number of inputted students is up to the user, therefore the program should ask whether the user want to input again or finish.
In the end of the program (after inputting information), it will show the number of the male and female students.

Must use: looping, conditional, array, struct.